chore(security): revoke other tokens for some ops
- Changing password
- Updating MFA
