+++ src/auth/index.ts
@@ -9,3 +9,5 @@ export * from "./webauthn";
 export * from "./organization";
 export * from "./roles";
 export * from "./saml";
+export * from "./pass";
+export * from "./otp";
